下载逻辑优化，修复下载闪退问题，修复查看曲谱可能的闪退问题，修复少数用户微博无法登录问题，增加分类功能，发布1.0.4

+++ app/src/main/java/com/yun/opern/common/WeiBoUserInfo.java
@@ -626,7 +626,7 @@ public class WeiBoUserInfo implements Serializable{
         private boolean isLongText;
         private int mlevel;
         private VisibleBean visible;
-        private int biz_feature;
+        private long biz_feature;
         private int hasActionTypeCard;
         private int userType;
         private int more_info_type;
@@ -824,11 +824,11 @@ public class WeiBoUserInfo implements Serializable{
             this.visible = visible;
         }
 
-        public int getBiz_feature() {
+        public long getBiz_feature() {
             return biz_feature;
         }
 
-        public void setBiz_feature(int biz_feature) {
+        public void setBiz_feature(long biz_feature) {
             this.biz_feature = biz_feature;
         }
